Move draw_primitive into p5_helpers
