+++ Coral/Maze/Maze.js
@@ -1,5 +1,5 @@
 import { GroupPrimitive } from "../../sketchlib/rendering/primitives.js";
-import { draw_primitive } from "../../sketchlib/draw_primitive.js";
+import { draw_primitive } from "../../sketchlib/p5_helpers/draw_primitive.js";
 import { generate_maze } from "../../sketchlib/RandomDFSMaze.js";
 import { CoralTile } from "../CoralTile.js";
 import { Rect } from "../Rect.js";
