+++ MosaicSlider/MosaicSlider.js
@@ -1,6 +1,6 @@
 import { WIDTH, HEIGHT } from "../sketchlib/dimensions.js";
 import { Color } from "../sketchlib/Color.js";
-import { draw_primitive } from "../sketchlib/draw_primitive.js";
+import { draw_primitive } from "../sketchlib/p5_helpers/draw_primitive.js";
 import { fix_mouse_coords } from "../sketchlib/fix_mouse_coords.js";
 import { InteractiveMosaic } from "./InteractiveMosaic.js";
 import { prevent_mobile_scroll } from "../sketchlib/prevent_mobile_scroll.js";
